Updating PingDataConnector with drop and delete data

+++ app/src/main/java/storage/mthomson/coneath/PingData.java
@@ -8,5 +8,5 @@ import java.sql.Date;
 
 public class PingData {
     public Date Timestamp;
-    public int PingValue;
+    public double PingValue;
 }
